+++ src/Model/YamlServiceAliasing/YamlServiceAliasingChecker.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace YamlStandards\Model\YamlServiceAliasing;
 
 use SebastianBergmann\Diff\Differ;
+use SebastianBergmann\Diff\Output\UnifiedDiffOutputBuilder;
 use YamlStandards\Model\AbstractChecker;
 use YamlStandards\Model\Component\YamlService;
 use YamlStandards\Model\Config\StandardParametersData;
@@ -35,7 +36,7 @@ class YamlServiceAliasingChecker extends AbstractChecker
             return new Result($pathToFile, Result::RESULT_CODE_OK);
         }
 
-        $differ = new Differ();
+        $differ = new Differ(new UnifiedDiffOutputBuilder());
         $diffBetweenStrings = $differ->diff($yamlContent, $correctYamlContent);
 
         return new Result($pathToFile, Result::RESULT_CODE_INVALID_FILE_SYNTAX, $diffBetweenStrings, true);